DP-1575 Add support for default task config values in pipeline
Add a new `taskConfig` field on the `Config` class which contains the
task configuration for the current state machine task.

Default task configuration values can be added to the pipeline template,
by adding a `task_config` field on `Parameters` of the pipeline step.
This can be overridden per pipeline by the `task_config` field on the
pipeline instance configuration supplied when creating the pipeline.

It also allows for missing (or `null`) task configurations, both on
pipeline template and instance level.

The implementation and test cases have been adapted from `common-python`
in attempt to match its behavior as closely as possible in this JVM
counterpart.